+++ wogu-temporal/src/test/java/dev/wogu/temporal/RuleDefinitionTest.java
@@ -27,6 +27,7 @@ class RuleDefinitionTest {
         List.of(),
         List.of(),
         null,
+        List.of(),
         List.of());
   }